requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pyyaml==6.0.2
requests==2.32.3
beautifulsoup4==4.12.3
markdownify==0.13.1
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
transformers==4.44.2
torch>=2.2.0
nltk==3.9.1
mlflow==2.14.3
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.36
apscheduler==3.10.4
httpx==0.27.2
pytest==8.1.1
python-slugify==8.0.4